+++ management-api-test/src/test/java/com/datastax/mgmtapi/BaseDockerIntegrationTest.java
@@ -107,13 +107,12 @@ public abstract class BaseDockerIntegrationTest {
   public static List<String> testVersions() {
     List<String> versions = new ArrayList<>(4);
 
-    if (Boolean.getBoolean("run3.11tests")) versions.add("3_11");
-    if (Boolean.getBoolean("run3.11testsUBI")) versions.add("3_11_ubi");
     if (Boolean.getBoolean("run4.0tests")) versions.add("4_0");
     if (Boolean.getBoolean("run4.0testsUBI")) versions.add("4_0_ubi");
     if (Boolean.getBoolean("run4.1tests")) versions.add("4_1");
     if (Boolean.getBoolean("run4.1testsUBI")) versions.add("4_1_ubi");
     if (Boolean.getBoolean("run5.0testsUBI")) versions.add("5_0_ubi");
+    if (Boolean.getBoolean("run6.0testsUBI")) versions.add("6_0_ubi");
     if (Boolean.getBoolean("runtrunktestsUBI")) versions.add("trunk_ubi");
     if (Boolean.getBoolean("runDSE6.8tests")) versions.add("dse-68");
     if (Boolean.getBoolean("runDSE6.8testsUBI")) versions.add("dse-68_ubi");
@@ -244,9 +243,6 @@ public abstract class BaseDockerIntegrationTest {
   }
 
   protected int getNumTokenRanges() {
-    if (this.version.startsWith("3")) {
-      return 256;
-    }
     if (this.version.startsWith("dse")) {
       return 1;
     }
@@ -254,6 +250,9 @@ public abstract class BaseDockerIntegrationTest {
       return 16;
     }
     if (this.version.startsWith("5")) {
+      return 16;
+    }
+    if (this.version.startsWith("6")) {
       return 16;
     }
     if (this.version.startsWith("trunk")) {